Rewrote acceptance test suite to use `behat/behat` as a main runner
While `psalm/codeception-psalm-module` is a good idea in theory,
in practice, Gherkin scenario steps as a composer dependency are really (REALLY)
hard to distribute and maintain together with a test suite.

This move replaces the external dependency with a local one, with the
test runner mostly staying out of the way, and test logic all implemented
in the `Context` class we wrote ourselves.

The main aim is to improve atomicity of changes on the test suite itself.

Ref: https://github.com/psalm/codeception-psalm-module/issues/54